+++ exchanges_wrapper/bitfinex_parser.py
@@ -124,6 +124,12 @@ def exchange_info(symbols_details: [], tickers: []) -> {}:
                 "applyToMarket": True,
                 "avgPriceMins": 0
             }
+            _percent_price = {
+                "filterType": "PERCENT_PRICE",
+                "multiplierUp": "5",
+                "multiplierDown": "0.2",
+                "avgPriceMins": 5
+            }
 
             symbol = {
                 "symbol": _symbol,
@@ -143,7 +149,7 @@ def exchange_info(symbols_details: [], tickers: []) -> {}:
                 "cancelReplaceAllowed": False,
                 "isSpotTradingAllowed": True,
                 "isMarginTradingAllowed": False,
-                "filters": [_price_filter, _lot_size, _min_notional],
+                "filters": [_price_filter, _lot_size, _min_notional, _percent_price],
                 "permissions": ["SPOT"],
             }
             symbols.append(symbol)